svg-sprite.js: Normalize `file` and `name` paths
This uses `path.normalize()` to ensure `file` and `name` values use the
same path separators.

This also tests if both variables are not the empty string before
calling `path.normalize()` as it returns `.` in that case.

`path.normalize()` also removes `./` segments so it is not necessary to
call `trimStart()` on `name`.

Fixes https://github.com/svg-sprite/svg-sprite/issues/942